GameView renamed to DisplayGame; Added empty lines before comments, after closing brackets; Displaying game data moved to DisplayGame class; New class DataManager for Save un Load functions; AliveCells renamed to WillCellBeAlive;
