+++ GameOfLife/Controller/GameManager.cs
@@ -1,6 +1,4 @@
-﻿using Newtonsoft.Json;
-
-namespace GameOfLife
+﻿namespace GameOfLife
 {
     /// <summary>
     /// Management of the game.
@@ -17,15 +15,12 @@ namespace GameOfLife
             int column = UserInterface.GetValidUserInput(GameMessages.inputColomnMessage, GameParameters.minInputValue, GameParameters.maxInputValue);
             int row = UserInterface.GetValidUserInput(GameMessages.inputRowMessage, GameParameters.minInputValue, GameParameters.maxInputValue);
             Game game = new(column, row);
-            GameView.View(game.GameField);
+            DisplayGame.Display(game.GameField, game);
 
             while (true)
             {
                 game.GetNextGeneration();
-                GameView.View(game.GameField);
-                Console.WriteLine($"Iteration: {game.CountOfIteration}");
-                Console.WriteLine($"Live cells: {game.CountOfLiveCells}");
-                Console.WriteLine(GameMessages.pauseGame);
+                DisplayGame.Display(game.GameField, game);
                 Thread.Sleep(1000);
 
                 if (Console.KeyAvailable && Console.ReadKey(true).Key == ConsoleKey.Spacebar)
@@ -34,11 +29,11 @@ namespace GameOfLife
                     string choice = Console.ReadLine().ToLower();
                     if (choice == "s")
                     {
-                        SaveGame(game);
+                        DataManager.SaveGame(game);
                     }
                     else if (choice == "l")
                     {
-                        game = LoadGame();
+                        game = DataManager.LoadGame();
                     }
                     else if (choice == "x")
                     {
@@ -46,25 +41,6 @@ namespace GameOfLife
                     }
                 }
             }
-        }
-
-        /// <summary>
-        /// Saves current game generation to the file.
-        /// </summary>
-        /// <param name="game">Current game generation.</param>
-        public static void SaveGame(Game game)
-        {
-            File.WriteAllText(GameMessages.fileName, JsonConvert.SerializeObject(game));
-        }
-
-        /// <summary>
-        /// Loads previously saved game file.
-        /// </summary>
-        public static Game? LoadGame()
-        {
-            Game loadgame = JsonConvert.DeserializeObject<Game>(File.ReadAllText(GameMessages.fileName));
-            return loadgame;
-
         }
     }
 }
