+++ GameOfLife/View/DisplayGame.cs
@@ -3,13 +3,13 @@
     /// <summary>
     /// View of the game field.
     /// </summary>
-    public class GameView
+    public class DisplayGame
     {
         /// <summary>
         /// Makes view of the game field.
         /// </summary>
         /// <param name="args">Array containing the game field.</param>
-        public static void View(int[,] args)
+        public static void Display(int[,] args, Game game)
         {
             Console.Clear();
             for (int row = 0; row < args.GetLength(1); row++)
@@ -18,8 +18,13 @@
                 {
                     Console.Write(args[column, row] == 1 ? "*" : " ");
                 }
+
                 Console.WriteLine();
             }
+
+            Console.WriteLine($"Iteration: {game.CountOfIteration}");
+            Console.WriteLine($"Live cells: {game.CountOfLiveCells}");
+            Console.WriteLine(GameMessages.pauseGame);
         }
     }
 }
